+++ azure-monitor-opentelemetry/azure/monitor/opentelemetry/util/configurations.py
@@ -9,6 +9,7 @@ from os import environ
 from typing import Dict
 
 from azure.monitor.opentelemetry._constants import (
+    DISABLE_AZURE_CORE_TRACING_ARG,
     DISABLE_LOGGING_ARG,
     DISABLE_METRICS_ARG,
     DISABLE_TRACING_ARG,
@@ -52,6 +53,7 @@ def _get_configurations(**kwargs) -> Dict[str, ConfigurationValue]:
     _default_disabled_instrumentations(configurations)
     _default_logging_export_interval_ms(configurations)
     _default_sampling_ratio(configurations)
+    _default_disable_azure_core_tracing(configurations)
 
     # TODO: remove when validation added to BLRP
     if configurations[LOGGING_EXPORT_INTERVAL_MS_ARG] <= 0:
@@ -125,3 +127,8 @@ def _default_sampling_ratio(configurations):
                 _INVALID_FLOAT_MESSAGE, SAMPLING_RATIO_ENV_VAR, default, e
             )
     configurations[SAMPLING_RATIO_ARG] = default
+
+
+# TODO: Placeholder for future configuration
+def _default_disable_azure_core_tracing(configurations):
+    configurations[DISABLE_AZURE_CORE_TRACING_ARG] = False